REPL compiles piecemeal. No more redundancy.
